BACKEND/requirements.txt
Django==4.1.1
djangorestframework==3.14.0
Pillow==9.2.0
djangorestframework-simplejwt==5.2.1
django-environ==0.9.0
django-cors-headers==3.13.0
